+++ packages/cloud_firestore/cloud_firestore_web/lib/src/transaction_web.dart
@@ -35,7 +35,7 @@ class TransactionWeb extends TransactionPlatform {
 
       return convertWebDocumentSnapshot(this._firestore, webDocumentSnapshot);
     } catch (e, s) {
-      throw getFirebaseException(e, s);
+      throw convertPlatformException(e, s);
     }
   }