+++ packages/cloud_firestore/cloud_firestore_web/lib/src/write_batch_web.dart
@@ -23,7 +23,7 @@ class WriteBatchWeb extends WriteBatchPlatform {
     try {
       await _webWriteBatchDelegate.commit();
     } catch (e, s) {
-      throw getFirebaseException(e, s);
+      throw convertPlatformException(e, s);
     }
   }