+++ packages/firebase_auth/firebase_auth/ios/Classes/FLTFirebaseAuthPlugin.m
@@ -247,7 +247,7 @@ NSString *const kErrMsgInvalidCredential =
 #pragma mark - AppDelegate
 
 #if TARGET_OS_IPHONE
-- (bool)application:(UIApplication *)application
+- (BOOL)application:(UIApplication *)application
     didReceiveRemoteNotification:(NSDictionary *)notification
           fetchCompletionHandler:(void (^)(UIBackgroundFetchResult result))completionHandler {
   if ([[FIRAuth auth] canHandleNotification:notification]) {
@@ -259,7 +259,7 @@ NSString *const kErrMsgInvalidCredential =
 
 - (void)application:(UIApplication *)application
     didRegisterForRemoteNotificationsWithDeviceToken:(NSData *)deviceToken {
-  [[FIRAuth auth] setAPNSToken:deviceToken type:FIRAuthAPNSTokenTypeProd];
+  [[FIRAuth auth] setAPNSToken:deviceToken type:FIRAuthAPNSTokenTypeUnknown];
 }
 
 - (BOOL)application:(UIApplication *)app openURL:(NSURL *)url options:(NSDictionary *)options {
@@ -603,11 +603,11 @@ NSString *const kErrMsgInvalidCredential =
                                   @{kArgumentToken : (id)tokenResult.token ?: [NSNull null]});
                             } else {
                               long expirationTimestamp =
-                                  (long)[tokenResult.expirationDate timeIntervalSince1970];
+                                  (long)[tokenResult.expirationDate timeIntervalSince1970] * 1000;
                               long authTimestamp =
-                                  (long)[tokenResult.authDate timeIntervalSince1970];
+                                  (long)[tokenResult.authDate timeIntervalSince1970] * 1000;
                               long issuedAtTimestamp =
-                                  (long)[tokenResult.issuedAtDate timeIntervalSince1970];
+                                  (long)[tokenResult.issuedAtDate timeIntervalSince1970] * 1000;
 
                               NSMutableDictionary *tokenData =
                                   [[NSMutableDictionary alloc] initWithDictionary:@{
@@ -705,13 +705,16 @@ NSString *const kErrMsgInvalidCredential =
     return;
   }
 
-  [currentUser sendEmailVerificationWithCompletion:^(NSError *_Nullable error) {
-    if (error != nil) {
-      result.error(nil, nil, nil, error);
-    } else {
-      result.success(nil);
-    }
-  }];
+  FIRActionCodeSettings *actionCodeSettings =
+      [self getFIRActionCodeSettingsFromArguments:arguments];
+  [currentUser sendEmailVerificationWithActionCodeSettings:actionCodeSettings
+                                                completion:^(NSError *_Nullable error) {
+                                                  if (error != nil) {
+                                                    result.error(nil, nil, nil, error);
+                                                  } else {
+                                                    result.success(nil);
+                                                  }
+                                                }];
 }
 
 - (void)userUnlink:(id)arguments withMethodCallResult:(FLTFirebaseMethodCallResult *)result {
@@ -911,7 +914,7 @@ NSString *const kErrMsgInvalidCredential =
   @synchronized(self->_authChangeListeners) {
     if (_authChangeListeners[auth.app.name] == nil) {
       _authChangeListeners[auth.app.name] =
-          [[FIRAuth auth] addAuthStateDidChangeListener:authStateChangeListener];
+          [auth addAuthStateDidChangeListener:authStateChangeListener];
     }
   }
 
@@ -927,7 +930,7 @@ NSString *const kErrMsgInvalidCredential =
   @synchronized(self->_idTokenChangeListeners) {
     if (_idTokenChangeListeners[auth.app.name] == nil) {
       _idTokenChangeListeners[auth.app.name] =
-          [[FIRAuth auth] addIDTokenDidChangeListener:idTokenChangeListener];
+          [auth addIDTokenDidChangeListener:idTokenChangeListener];
     }
   }
 
@@ -1110,10 +1113,18 @@ NSString *const kErrMsgInvalidCredential =
   }
 
   NSString *signInMethod = credentialDictionary[kArgumentSignInMethod];
-  NSString *secret = credentialDictionary[kArgumentSecret];
-  NSString *idToken = credentialDictionary[kArgumentIdToken];
-  NSString *accessToken = credentialDictionary[kArgumentAccessToken];
-  NSString *rawNonce = credentialDictionary[kArgumentRawNonce];
+  NSString *secret = credentialDictionary[kArgumentSecret] == [NSNull null]
+                         ? nil
+                         : credentialDictionary[kArgumentSecret];
+  NSString *idToken = credentialDictionary[kArgumentIdToken] == [NSNull null]
+                          ? nil
+                          : credentialDictionary[kArgumentIdToken];
+  NSString *accessToken = credentialDictionary[kArgumentAccessToken] == [NSNull null]
+                              ? nil
+                              : credentialDictionary[kArgumentAccessToken];
+  NSString *rawNonce = credentialDictionary[kArgumentRawNonce] == [NSNull null]
+                           ? nil
+                           : credentialDictionary[kArgumentRawNonce];
 
   // Password Auth
   if ([signInMethod isEqualToString:kSignInMethodPassword]) {