+++ packages/firebase_auth/firebase_auth_web/lib/firebase_auth_web_recaptcha_verifier_factory.dart
@@ -3,10 +3,13 @@
 // found in the LICENSE file.
 
 import 'dart:async';
+import 'dart:html';
 
 import 'package:firebase/firebase.dart' as firebase;
 import 'package:firebase_auth_platform_interface/firebase_auth_platform_interface.dart';
 import 'package:firebase_auth_web/utils.dart';
+
+const String _kInvisibleElementId = '__ff-recaptcha-container';
 
 /// The delegate implementation for [RecaptchaVerifierFactoryPlatform].
 ///
@@ -27,21 +30,83 @@ class RecaptchaVerifierFactoryWeb extends RecaptchaVerifierFactoryPlatform {
   RecaptchaVerifierFactoryWeb._() : super();
 
   /// Creates a new [RecaptchaVerifierFactoryWeb] with a container and parameters.
-  RecaptchaVerifierFactoryWeb(
-      {String container, Map<String, dynamic> parameters})
-      : _delegate = firebase.RecaptchaVerifier(container, parameters),
-        super();
+  RecaptchaVerifierFactoryWeb({
+    String container,
+    RecaptchaVerifierSize size = RecaptchaVerifierSize.normal,
+    RecaptchaVerifierTheme theme = RecaptchaVerifierTheme.light,
+    RecaptchaVerifierOnSuccess onSuccess,
+    RecaptchaVerifierOnError onError,
+    RecaptchaVerifierOnExpired onExpired,
+  }) : super() {
+    String element;
+    Map<String, dynamic> parameters = {};
 
-  @override
-  RecaptchaVerifierFactoryPlatform delegateFor(
-      {String container, Map<String, dynamic> parameters}) {
-    return RecaptchaVerifierFactoryWeb(
-        container: container, parameters: parameters);
+    if (onSuccess != null) {
+      parameters['callback'] = (resp) {
+        onSuccess();
+      };
+    }
+
+    if (onExpired != null) {
+      parameters['expired-callback'] = () {
+        onExpired();
+      };
+    }
+
+    if (onError != null) {
+      parameters['error-callback'] = (Object error) {
+        onError(throwFirebaseAuthException(error));
+      };
+    }
+
+    if (container == null || container.isEmpty) {
+      parameters['size'] = 'invisible';
+      Element el = window.document.getElementById(_kInvisibleElementId);
+
+      // If an existing element exists, something may have already been rendered.
+      if (el != null) {
+        el.remove();
+      }
+
+      window.document.documentElement.children
+          .add(DivElement()..id = _kInvisibleElementId);
+
+      element = _kInvisibleElementId;
+    } else {
+      parameters['size'] = convertRecaptchaVerifierSize(size);
+      parameters['theme'] = convertRecaptchaVerifierTheme(theme);
+
+      Element el = window.document.getElementById(container);
+      assert(el != null,
+          'An exception was thrown whilst creating a RecaptchaVerifier instance. No DOM element with an ID of $container could be found.');
+
+      // If the provided string container ID has been found, assign it.
+      element = container;
+    }
+
+    _delegate = firebase.RecaptchaVerifier(element, parameters);
   }
 
   @override
-  T getDelegate<T>() {
-    return _delegate as T;
+  RecaptchaVerifierFactoryPlatform delegateFor(
+      {String container,
+      RecaptchaVerifierSize size = RecaptchaVerifierSize.normal,
+      RecaptchaVerifierTheme theme = RecaptchaVerifierTheme.light,
+      RecaptchaVerifierOnSuccess onSuccess,
+      RecaptchaVerifierOnError onError,
+      RecaptchaVerifierOnExpired onExpired}) {
+    return RecaptchaVerifierFactoryWeb(
+        container: container,
+        size: size,
+        theme: theme,
+        onSuccess: onSuccess,
+        onError: onError,
+        onExpired: onExpired);
+  }
+
+  @override
+  firebase.ApplicationVerifier get delegate {
+    return _delegate;
   }
 
   @override
@@ -62,9 +127,9 @@ class RecaptchaVerifierFactoryWeb extends RecaptchaVerifierFactoryPlatform {
   }
 
   @override
-  Future<int> render() {
+  Future<int> render() async {
     try {
-      return _delegate.render();
+      return (await _delegate.render()) as int;
     } catch (e, s) {
       throw throwFirebaseAuthException(e, s);
     }