+++ packages/cloud_functions/cloud_functions_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.0.0-dev.2
+
+ - Update a dependency to the latest release.
+
+## 3.0.0-dev.1
+
+* Refactor to introduce `HttpsCallablePlatform`, `HttpsCallableOptions` & `FirebaseFunctionsException` classes.
+
 ## 2.0.0
 
 * Update to depend on `firebase_core` plugin.