+++ packages/firebase_dynamic_links/CHANGELOG.md
@@ -1,6 +1,15 @@
-## 0.5.4-dev.1
+## 0.6.0+2
 
-* Depend on `firebase_core` pre-release.
+ - Update a dependency to the latest release.
+
+## 0.6.0+1
+
+ - **FIX**: local dependencies in example apps (#3319).
+ - **CHORE**: intellij cleanup (#3326).
+
+## 0.6.0
+
+* Depend on new `firebase_core` plugin.
 * Firebase iOS SDK versions are now locked to use the same version defined in
   `firebase_core`.
 * Firebase Android SDK versions are now using the Firebase Bill of Materials (BoM)